# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Payment Processing Backend API
  description: |
    RESTful API for payment processing with Authorize.Net integration.
    
    This API provides endpoints for:
    - User authentication and authorization (JWT-based)
    - Payment operations (purchase, authorize, capture, void, refund)
    - Order management
    
    All payment endpoints require JWT authentication. Use the `/auth/login` or `/auth/register` endpoint to obtain an access token.
  version: 1.0.0
  contact:
    name: Payment Processing Backend
    email: support@example.com

servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server

tags:
  - name: Authentication
    description: User authentication and token management
  - name: Payments
    description: Payment processing operations

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account and returns JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: SecurePassword123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                tokenType: Bearer
                userId: 1
                email: user@example.com
                role: USER
        '400':
          description: Validation error or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticates user and returns JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Uses refresh token to obtain a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Revokes all refresh tokens for the authenticated user
      responses:
        '204':
          description: Logout successful
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/purchase:
    post:
      tags:
        - Payments
      summary: Process a purchase (authorize + capture)
      description: |
        Processes a payment transaction by authorizing and capturing funds in a single step.
        This is the simplest payment flow for immediate payment processing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            example:
              orderId: 1
              amount: 100.00
              currency: USD
              paymentMethod:
                cardNumber: "4111111111111111"
                expiryMonth: "12"
                expiryYear: "2025"
                cvv: "123"
                cardholderName: John Doe
              description: Purchase order #12345
      headers:
        X-Idempotency-Key:
          description: Unique key for idempotency (UUID recommended)
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Purchase successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Validation error or bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Payment declined by gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/authorize:
    post:
      tags:
        - Payments
      summary: Authorize payment (hold funds)
      description: |
        Authorizes funds on a payment method without capturing them immediately.
        Funds are held but not captured. Use this for pre-authorization scenarios.
        Authorizations typically expire after 30 days.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            example:
              orderId: 1
              amount: 150.00
              currency: USD
              paymentMethod:
                cardNumber: "4111111111111111"
                expiryMonth: "12"
                expiryYear: "2025"
                cvv: "123"
                cardholderName: John Doe
              description: Authorization for order #12346
      headers:
        X-Idempotency-Key:
          description: Unique key for idempotency (UUID recommended)
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Authorization successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Authorization declined by gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/capture:
    post:
      tags:
        - Payments
      summary: Capture authorized funds
      description: |
        Captures funds that were previously authorized. Supports full or partial capture.
        The capture amount cannot exceed the remaining authorized amount.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
            example:
              authorizationTransactionId: 2
              amount: 150.00
              currency: USD
      headers:
        X-Idempotency-Key:
          description: Unique key for idempotency (UUID recommended)
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Capture successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Capture declined by gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Authorization transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule validation failed (e.g., authorization expired, amount exceeds remaining)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/cancel:
    post:
      tags:
        - Payments
      summary: Cancel/void authorization
      description: |
        Cancels an authorization before it is captured. Releases held funds.
        Must be done within 24 hours of authorization (Authorize.Net limitation).
        Cannot cancel an authorization that has already been captured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
            example:
              authorizationTransactionId: 2
      headers:
        X-Idempotency-Key:
          description: Unique key for idempotency (UUID recommended)
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Cancel/void successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Authorization transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule validation failed (e.g., already captured, outside void window)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/refund:
    post:
      tags:
        - Payments
      summary: Refund a transaction
      description: |
        Refunds funds from a previously captured or purchased transaction.
        Supports full and partial refunds. Multiple partial refunds are allowed.
        Refund amount cannot exceed the remaining refundable amount.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
            example:
              transactionId: 1
              amount: 50.00
              currency: USD
              reason: Customer requested refund
      headers:
        X-Idempotency-Key:
          description: Unique key for idempotency (UUID recommended)
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Refund successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Refund declined by gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule validation failed (e.g., amount exceeds refundable amount)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication using Bearer token.
        Include the token in the Authorization header as: `Authorization: Bearer <token>`
        Obtain a token by calling `/auth/login` or `/auth/register`.

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: User password (minimum 8 characters)
          example: SecurePassword123

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User password
          example: SecurePassword123

    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token obtained from login or register
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (expires in 15 minutes)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT refresh token (expires in 7 days)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          description: Token type (always "Bearer")
          example: Bearer
        userId:
          type: integer
          format: int64
          description: User ID
          example: 1
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        role:
          type: string
          enum:
            - USER
            - ADMIN
          description: User role
          example: USER

    PaymentRequest:
      type: object
      required:
        - orderId
        - amount
        - currency
        - paymentMethod
      properties:
        orderId:
          type: integer
          format: int64
          description: Order ID to process payment for
          example: 1
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Payment amount (must match order total for purchase/authorize)
          example: 100.00
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: Currency code (ISO 4217, e.g., USD)
          example: USD
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodRequest'
        description:
          type: string
          description: Optional payment description
          example: Purchase order #12345
        idempotencyKey:
          type: string
          format: uuid
          description: Optional idempotency key to prevent duplicate transactions
          example: 550e8400-e29b-41d4-a716-446655440000

    PaymentMethodRequest:
      type: object
      required:
        - cardNumber
        - expiryMonth
        - expiryYear
        - cvv
        - cardholderName
      properties:
        cardNumber:
          type: string
          pattern: '^\d{13,19}$'
          description: Credit card number (13-19 digits)
          example: "4111111111111111"
        expiryMonth:
          type: string
          pattern: '^(0[1-9]|1[0-2])$'
          description: Card expiry month (01-12)
          example: "12"
        expiryYear:
          type: string
          pattern: '^\d{4}$'
          description: Card expiry year (4 digits)
          example: "2025"
        cvv:
          type: string
          pattern: '^\d{3,4}$'
          description: Card CVV (3-4 digits)
          example: "123"
        cardholderName:
          type: string
          description: Cardholder name
          example: John Doe

    CaptureRequest:
      type: object
      required:
        - authorizationTransactionId
        - amount
        - currency
      properties:
        authorizationTransactionId:
          type: integer
          format: int64
          description: ID of the authorization transaction to capture
          example: 2
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Amount to capture (cannot exceed remaining authorized amount)
          example: 150.00
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: Currency code (must match authorization currency)
          example: USD
        idempotencyKey:
          type: string
          format: uuid
          description: Optional idempotency key
          example: 550e8400-e29b-41d4-a716-446655440000

    CancelRequest:
      type: object
      required:
        - authorizationTransactionId
      properties:
        authorizationTransactionId:
          type: integer
          format: int64
          description: ID of the authorization transaction to cancel/void
          example: 2
        reason:
          type: string
          description: Optional cancellation reason
          example: Customer requested cancellation

    RefundRequest:
      type: object
      required:
        - transactionId
        - amount
        - currency
      properties:
        transactionId:
          type: integer
          format: int64
          description: ID of the original PURCHASE or CAPTURE transaction to refund
          example: 1
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Refund amount (cannot exceed remaining refundable amount)
          example: 50.00
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: Currency code (must match original transaction currency)
          example: USD
        reason:
          type: string
          description: Optional refund reason
          example: Customer requested refund
        idempotencyKey:
          type: string
          format: uuid
          description: Optional idempotency key
          example: 550e8400-e29b-41d4-a716-446655440000

    PaymentResponse:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
          description: Internal transaction ID
          example: 1
        orderId:
          type: integer
          format: int64
          description: Order ID
          example: 1
        transactionType:
          type: string
          enum:
            - PURCHASE
            - AUTHORIZE
            - CAPTURE
            - VOID
            - REFUND
          description: Type of transaction
          example: PURCHASE
        status:
          type: string
          enum:
            - PENDING
            - SUCCESS
            - FAILED
            - VOIDED
            - REFUNDED
          description: Transaction status
          example: SUCCESS
        amount:
          type: number
          format: decimal
          description: Transaction amount
          example: 100.00
        currency:
          type: string
          description: Currency code
          example: USD
        gatewayTransactionId:
          type: string
          description: Authorize.Net transaction ID
          example: "60000000001"
        authorizationCode:
          type: string
          description: Authorization code (for authorize/purchase transactions)
          example: ABC123
        lastFourDigits:
          type: string
          description: Last 4 digits of card number
          example: "1111"
        cardBrand:
          type: string
          description: Card brand (e.g., Visa, MasterCard)
          example: Visa
        createdAt:
          type: string
          format: date-time
          description: Transaction creation timestamp
          example: "2024-01-15T10:30:00"
        order:
          $ref: '#/components/schemas/OrderSummary'

    OrderSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Order ID
          example: 1
        orderNumber:
          type: string
          description: Order number
          example: ORD-001
        status:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - COMPLETED
            - PARTIALLY_CAPTURED
            - PARTIALLY_REFUNDED
            - REFUNDED
            - CANCELLED
            - FAILED
          description: Order status
          example: COMPLETED
        totalAmount:
          type: number
          format: decimal
          description: Order total amount
          example: 100.00

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Error message
          example: Validation failed
        path:
          type: string
          description: API path where error occurred
          example: /api/payments/purchase
        errors:
          type: object
          additionalProperties:
            type: string
          description: Validation errors (field -> error message mapping)
          example:
            email: Email is required
            password: Password must be at least 8 characters
        gatewayErrorCode:
          type: string
          description: Payment gateway error code (if applicable)
          example: "5"
        gatewayErrorMessage:
          type: string
          description: Payment gateway error message (if applicable)
          example: Do Not Honor
        transactionId:
          type: integer
          format: int64
          description: Transaction ID related to error (if applicable)
        existingTransactionId:
          type: integer
          format: int64
          description: Existing transaction ID for duplicate transaction errors
        currentStatus:
          type: string
          description: Current status when status validation fails
        requiredStatus:
          type: string
          description: Required status when status validation fails
        allowedOperations:
          type: string
          description: Allowed operations when operation validation fails
        authorizationExpiresAt:
          type: string
          format: date-time
          description: Authorization expiration timestamp (if applicable)
        authorizedAmount:
          type: number
          format: decimal
          description: Authorized amount (if applicable)
        capturedAmount:
          type: number
          format: decimal
          description: Captured amount (if applicable)
        refundedAmount:
          type: number
          format: decimal
          description: Refunded amount (if applicable)